Cancel touchscreen gestures from Dispatcher when display rotates
When the display rotates, previously, touchscreen devices were reset,
and a notifyDeviceReset is sent to Displatcher. When that happens,
Dispatcher cancells all active gestures from that device.

With this CL, we skip a device reset when the display rotates because
there is nothing that changes for the touchscreen device. Instead, when
we get a new set of input windows from SF in dispatcher, we track if the
orientation changes, and if it does, issue a cancellation for all
pointer events.

Bug: 185943742
Test: manual
Change-Id: Ic8ff19c0dc9cca7b1053a137ea041e5a0da82a76