Add a stricter check for resetting all cached sets.
Presently cached sets are reset if the previously seen layer geometry
hashed to the same value as the incoming layer geometry. But because
hashing does not guarantee uniqueness, this check is not inclusive
enough to check for invalidating the cached sets since it does not
correctly invalidate when there is a colliding geometry.

Instead, check for equality of the LayerState geometries directly, which
is computationally equivalent to computing the hash but guarantees that
unique geometries do not equal each other.

Bug: 185728087
Test: libcompositionengine_test
Test: camera app no longer crashes when switching between cameras.
Change-Id: I2a5e918464690d7e7f54ecd73b8eb7631d2aed9c