[SF] When content detection is off layer should have NoVote
Currently when content detection is OFF, layer vote with Max,
which interferes with layers with explicit vote coming from
setFrameRate. In this CL we change the default for such
layers to be NoVote.
Note that if all visible layers have a NoVote then RefreshRateConfigs
will still select the Max refresh rate.

Bug: 183128900
Test: atest RefreshRateConfigs SchedulerTest
Change-Id: Ia7a023ed9926371de225ab94ce0d46c6bbf0479e