Ensure that blurRegions take into account the buffers rotation.
Some tranformations are applied to content rendered into a buffer that
are unknown to the UI (e.g. Android View) that drew it. That is the case
for blurRegions so we must provide a matrix that translates the contents
of the blurRegion into the same coordinate space used by the other
geometry in RenderEngine's LayerSettings.

Bug: 185228719
Test: atest librenderengine_test and SilkFX demo
Change-Id: I4e3144bac678ffff2f29f3270d589e9f3fe94634