Do not map protected buffers to GPU resources and cache the result
This is in line with the older version of RenderEngine and prevents
the device from exhausting its limited protected GPU memory on cached
objects.

This CL also ensures that upon switching to/from a protected context
that the Skia GPU state is properly initialized and any scratch
resources that were being held onto are purged.

Test: atest com.google.android.media.gts.WidevineDashPolicyTests
Bug: 186158883
Bug: 183391755
Change-Id: I31c03a7589feea6f5d2eb244f77b934e1868326d