Updated SkiaGLRenderEngine to no longer adjust stretch bounds
Because the stretch region for surfaces is now calculated
within HWUI as part of the transformation matrix, remove
the position logic to adjust the drawing bounds for a
surface with stretch applied to it. Also disabled
the stretch shader within render engine as
the approximate stretch from the matrix transformation
provides reasonable results.

For consistency update the shader within SkiaGLRenderEngine
to match easing changes from cubic to linear.

Bug: 179047472
Test: manual

Change-Id: I66170ba4e3168186b3d3e4e27ef6a26b4637bc18