SurfaceFlinger: previousFrameFence needs to check the current time
previousFrameFence returns the fence associated with the previous vsync.
In case that SF wake up was delayed due to a previous frame, up to the
point that the wake up occurs after the vsync, SF needs to check fence
N-1 and not N-2 since frame N-1 was targeting the vsync prior to SF
wake up.

Bug: 187886038
Test: run bouncy ball and simulate the first frame missed
Change-Id: I176c6641dd42132fcfd75a79898a5d2bd8da9c90