Fix shadow corner radius during PIP size change
Bug: 187441996
Test: manual
Test: go/wm-smoke
Test: libcompositionengine_test
Test: TODO

The now-deleted code was intended to remove rounding when part of the
content is offscreen, but it removes rounding on corners that are
entirely on screen. Remove the code, which fixes a bug seen when
expanding a PIP. If the content is partially offscreen, our drawing code
will handle this correctly.

In addition, if the bounds is an SkRect and there is a roundRectClip,
draw the rounded version.

Change-Id: I6ee07321c04474f96a572015f301e01441ca5afc