InputReader_tests: Fix tests to work with per-window-input-rotation
The per-window-input-rotation flag makes InputReader work in the
un-rotated coordinate space. This means rotation affects input devices
in the opposite way as without the flag. Devices that were previously
affected by rotation (orientation-aware devices) are no longer affected,
and vice versa.

We update the tests to take this into account and test the case when the
flag is enabled.

Bug: 179274888
Test: atest inputflinger_tests
Change-Id: I9901cc065409b7d414d09d77f8f8e9f2a99994e4