[BugFix] Reset key repeating in case a keyboard device was disabled or enabled.
The "resetKeyRepeatLocked" will never be called when we disable a keyboard device. In this case, if we long click a key on keyboard and it began to repeat. Then we disable the keyboard. We find that the key repeating will never stop until a next key event come.

Test: atest inputflinger_tests
Signed-off-by: liushenxiang <liushenxiang@xiaomi.com>
Change-Id: Ic0e37d1e1e447d4965f3fbb479e0eca7c726a6ab
(cherry picked from commit 97b0d3b5fcbe4dcf0a8d359f52027bbf49a8f387)