SF: Recreate display on resolution changes
When the resolution changes from SurfaceFlinger we
need to recreate the frambuffer. This CL reuses the
existing mechanism for processing display changes
to do that. More specifically the logic in this CL
does the same as if there was a subsequent hotplug
"connected" event.

This removes the need HWC HAL implementations to
send hotplug events on each resolution change
(which was violating VTS).

This change impacts only devices which support
multiple resolutions (e.g. TV but no phones).

Test: atest DisplayTest SetFrameRateTest
Test: plug hdmi out and in and verify that there are
      no crashes
Bug: 188684566
Change-Id: I8a00657e9eeb1db716ce58aac7b3e3e40f898352