Temporarily turn off latency reporting
The latency tracker is currently causing a crash due to a failed
assumption. Inside the tracker, there are 2 collections that are
expected to always be in sync. However, in dogfood, we are seeing:
Abort message: 'Event 102634453 is in mEventTimes, but not in mTimelines'

While we are trying to identify the issue, let's turn off the latency
reporting by simply not invoking the 'trackListener' portion of the
latency tracker.

This will avoid calling 'reportAndPruneMatureRecords', which is
currently the source of the crash.
This will not negatively impact the calls to 'trackFinishedEvent' and
'trackGraphicsLatency' because those will just quit early if the event
is not found.

Bug: 189319094
Test: presubmit
Change-Id: I3282eb30622d48f13a380020704105981bd3ae39