TouchVideoDevice: Silence log spam
When it was originally written, the TouchVideoDevice could expect that
every queued heatmap frame would correspond directly to a set of touch
coordinates. This is no longer true because touch coordinates may be
dropped before reaching InputFlinger. Therefore, heatmap frames may be
dropped frequently and the accompanying error message is less valuable.

Bug: 184000818
Bug: 189901803
Change-Id: I9ba848c70c0c5014e6ed5c2af8fd0f5629971254