Remove fatal assert from CachedSet::render
Bug: 189392914
Test: libcompositionengine_test
Test: manual

The intent was to catch a case where we somehow were looking at the
wrong layer (e.g. a shadow instead of the PIP itself). In practice, the
PIP will only have one layer. That said, if the content was protected,
BufferLayer::prepareClientComposition will not have supplied the buffer;
instead, it will have called prepareClearClientComposition to draw pure
black instead (since the CachedSet does not support protected content).
The intent of the CachedSet is to draw transparent anyway, so it is fine
that the buffer is not included in the LayerSettings.

Change-Id: I6a69886cca6f78b5f704c78457ee50f381b1c49d