eglSurfaceAttrib EGL_TIMESTAMPS_ANDROID non-window surface fix.
The spec says:
"If attribute is EGL_TIMESTAMPS_ANDROID, then values specifies whether
to enable/disable timestamp collection for this surface. A value of
EGL_TRUE enables timestamp collection, while a value of EGL_FALSE
disables it. The initial value is false. If surface is not a window
surface this has no effect."

Setting an error is considered a side effect, so return
EGL_TRUE for EGL_TIMESTAMPS_ANDROID attribute.

BUG: 174481138
Test: APK b/174481138#comment1, `adb logcat *:E -e egl` no error
Change-Id: Ia6f04126bd3dead47648a1c71f7100224f28553d