CachedSet::append: remove mHolePunchLayer
Bug: 189899692
Test: libcompositionengine_test

When the PIP is paused, it soon becomes inactive, and then it is
appended to the CachedSet behind it. This invalidates the override
buffer, but when we create a new one, we drew the video and (because
mHolePunchLayer was unchanged) immediately erased it.

With this CL, we'll remove mHolePunchLayer when we append so that we no
longer erase the PIP.

Change-Id: I45f4f8d557da9b54680527b08df23475ae04d0bc