Coalesce handlePageFlip and doTransaction traversal
Previous moved code had to execute either before commitTransaction or
update input-flinger. The remaining code in the doTransaction traversal
can execute later (around handlePageFlip). There is still room
for optimization here, in just removing this code, but it will
take more time, and the runtime is pretty fast at this point
once we coalesce the traversal.

Test: Existing tests pass. simpleperf
Bug: 186200583
Change-Id: I53d64db0c17df759904bbb6c1581cc2603a67a48