Switch away from std::async for proximity-active callback.
By default, std::async() returns a future with the expectation that the
caller monitors it. By ignoring the future we forced the future's
destructor to run which waits on the thread by default (causing a
deadlock). Since we don't really want to monitor the future for our
purposes here anyway, we're switching to a detached thread with
this change.

Test: Verify that when sensor is enabled, there is an additional
      refresh rate vote for the rate specified in config file.
Bug: 175793106
Change-Id: Icfc2fc7e8dc84c614a9a19d7f23c9ce6d55cc6f0