swapchain: increase the minimal buffer count to 3
On Android, the compositor will latch buffers only if they are completed
(i.e. the gpu fence is signaled), which means 2 buffers would
be acquired from the buffer queue. If the client wants mailbox or
swap_interval=0 behavior, an additional buffer is needed for the app
to never block on dequeueBuffer.

Test: run Vk cube and observe buffer count
Test: CtsDeqpTestCases: dEQP-VK.wsi.*
Bug: 182887411
Change-Id: I4c6bcb85200c6646dd8c504b2d40471c7d6eeed6